WIP on integration. Single parts seem to work fine with simulator, but it completely crashes if combined. We need to find a method to determine the initial movement direction.Also, both the particle filter and the poisson-solver seem to be way to slow to enable efficient testing.

+++ source/camera_simulation/cameraToWorldCoordinates.m
@@ -1,6 +1,6 @@
 function [ world_coordinates ] = cameraToWorldCoordinates( u, v, K, theta, img_size )
 
-invKPs = reshape(K \ [u v 1]', 1, 1, 3);
+invKPs = reshape(K \ double([u v 1]'), 1, 1, 3);
 world_coordinates = cameraToWorldCoordinatesBatch(invKPs(:,:,1:2), theta, img_size);
 
 end
